I added a very experimental feature today. Right now it cannot be turned off since I am looking for feedback. Once I've gotten some feedback it will be either removed or made toggle-able. I don't think it will trouble people too unduely for now.
Now, if you are not in your home clan ChIT will display your current clan name at the top of the charpane. You can change your home clan with:
 zlib chit.clan.home = clan name

Eventually chit.clan.display may have options for on, off and away. (Where "away" refers to the current behavior of only displaying when you are away from home.)

I may or may not implement a whitelist picker to drop-down from clan name when it is on.