I understand that you can learn skills from multiple classes in Avatar of West of Loathing. You get the books for ALL the classes you completed.
This means I'd either have to do away with a simple guild link as meaningless or create a whole new place to learn skills. I chose the later option because I'm a glutton for punishment and I recently invented such a technique.

I haven't decided if I am going to add any additional information or aesthetics to the Bookshelf of the West.